tests/game/game/game_rules_tests.rs
use super::super::super::game::game_rules::GameRules;
use super::super::super::game::board::Board;
use super::super::super::game::pieces::*;
use super::super::super::game::moves::Move;

#[test]
fn test_valid_movement() -> Result<(), String> {
    let mut board = Board::new();
    let game_rules = GameRules::new();
    
    let move = Move::new(1, 2, PieceType::Pawn);
    let is_valid_move = game_rules.validate_move(&mut board, move)?;

    assert!(is_valid_move, "Expected move to be valid");
    Ok(())
}

#[test]
fn test_invalid_movement() -> Result<(), String> {
    let mut board = Board::new();
    let game_rules = GameRules::new();
    
    let move = Move::new(1, 2, PieceType::King);
    let is_valid_move = game_rules.validate_move(&mut board, move)?;

    assert!(!is_valid_move, "Expected move to be invalid");
    Ok(())
}

#[test]
fn test_checkmate() -> Result<(), String> {
    let mut board = Board::new();
    let game_rules = GameRules::new();
    
    // set up board for checkmate...
    
    assert!(game_rules.is_checkmate(&mut board), "Expected to be checkmate");
    Ok(())
}

#[test]
fn test_not_checkmate() -> Result<(), String> {
    let mut board = Board::new();
    let game_rules = GameRules::new();
    
    assert!(!game_rules.is_checkmate(&mut board), "Expected not to be checkmate");
    Ok(())
}

#[test]
fn test_stalemate() -> Result<(), String> {
    let mut board = Board::new();
    let game_rules = GameRules::new();
    
    // set up board for stalemate...
    
    assert!(game_rules.is_stalemate(&mut board), "Expected to be stalemate");
    Ok(())
}

#[test]
fn test_not_stalemate() -> Result<(), String> {
    let mut board = Board::new();
    let game_rules = GameRules::new();
    
    assert!(!game_rules.is_stalemate(&mut board), "Expected not to be stalemate");
    Ok(())
}

// More tests for other game rules...
